refactor(quotes): use strong random number generator for trade sampler (#205)

+++ pkg/quotes/config.go
@@ -135,6 +135,6 @@ func (SyncswapConfig) DriverType() DriverType {
 }
 
 type TradeSamplerConfig struct {
-	Enabled           bool `yaml:"enabled" env:"QUOTES_TRADE_SAMPLER_ENABLED"`
-	DefaultPercentage int  `yaml:"default_percentage" env:"QUOTES_TRADE_SAMPLER_DEFAULT_PERCENTAGE"`
+	Enabled           bool  `yaml:"enabled" env:"QUOTES_TRADE_SAMPLER_ENABLED"`
+	DefaultPercentage int64 `yaml:"default_percentage" env:"QUOTES_TRADE_SAMPLER_DEFAULT_PERCENTAGE"`
 }
